+++ app/components/omnibox/services/search-service.js
@@ -148,7 +148,7 @@ angular.module('omnibox')
     this.zoomToSearchResult = function (result, state) {
       var ZOOM_FOR_OBJECT = 19;
 
-      if (state.box.type !== 'multi-point') {
+      if (state.box.type !== 'multi-point' && state.context !== 'dashboard') {
         state.selected.reset();
       }